bot reporting implemented
